Vector2unorm16
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1210 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/source/ImageFormat_convert.cpp
@@ -817,7 +817,7 @@ static unorm8 applyFilter(const unorm8*    I,
     
 	for (int dx = 0; dx < 5; ++dx) {
 	    float f = filter[dy][dx];
-	    sum += f * I[((x + dx + w - 2) % w) + offset];
+	    sum += f * (float)I[((x + dx + w - 2) % w) + offset];
 	    denom += f;
 	}
     }
